+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/application/Configurable.kt
@@ -1,10 +1,11 @@
 package com.insyncwithfoo.pyrightls.configuration.application
 
 import com.insyncwithfoo.pyrightls.configuration.common.PyrightLSConfigurable
+import com.insyncwithfoo.pyrightls.message
 import com.intellij.util.xmlb.XmlSerializerUtil
 
 
-class PyrightLSApplicationConfigurable : PyrightLSConfigurable<Configurations>() {
+internal class PyrightLSApplicationConfigurable : PyrightLSConfigurable<Configurations>() {
     
     override val service = ConfigurationService.getInstance()
     override val panel by lazy { ConfigurationPanel() }
@@ -12,10 +13,6 @@ class PyrightLSApplicationConfigurable : PyrightLSConfigurable<Configurations>()
     override val originalConfigurations: Configurations =
         XmlSerializerUtil.createCopy(service.configurations)
     
-    override fun getDisplayName() = DISPLAY_NAME
-    
-    companion object {
-        const val DISPLAY_NAME = "Pyright LS (Global)"
-    }
+    override fun getDisplayName() = message("configurations.global.displayName")
     
 }