+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/project/Configurable.kt
@@ -1,11 +1,13 @@
 package com.insyncwithfoo.pyrightls.configuration.project
 
 import com.insyncwithfoo.pyrightls.configuration.common.PyrightLSConfigurable
+import com.insyncwithfoo.pyrightls.message
 import com.intellij.openapi.project.Project
 import com.intellij.util.xmlb.XmlSerializerUtil
 
 
-class PyrightLSProjectConfigurable internal constructor(project: Project) : PyrightLSConfigurable<Configurations>() {
+internal class PyrightLSProjectConfigurable internal constructor(project: Project) :
+    PyrightLSConfigurable<Configurations>() {
     
     override val service = ConfigurationService.getInstance(project)
     override val panel by lazy { ConfigurationPanel(project) }
@@ -13,10 +15,6 @@ class PyrightLSProjectConfigurable internal constructor(project: Project) : Pyri
     override val originalConfigurations: Configurations =
         XmlSerializerUtil.createCopy(service.configurations)
     
-    override fun getDisplayName() = DISPLAY_NAME
-    
-    companion object {
-        const val DISPLAY_NAME = "Pyright LS (Project)"
-    }
+    override fun getDisplayName() = message("configurations.project.displayName")
     
 }