+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/project/ConfigurationService.kt
@@ -12,7 +12,7 @@ import com.intellij.util.xmlb.XmlSerializerUtil
 
 @State(name = "ProjectConfigurations", storages = [Storage("pyright-langserver.xml")])
 @Service(Service.Level.PROJECT)
-class ConfigurationService :
+internal class ConfigurationService :
     SimplePersistentStateComponent<Configurations>(Configurations()),
     HasConfigurations<Configurations> {