+++ CHANGELOG.md
@@ -5,16 +5,27 @@
 
 ## [Unreleased]
 
+### Added
+
+* The plugin will now suggest setting an executable
+  for the current project if one can be found locally.
+  To turn off suggestion for a project, uncheck
+  the corresponding option in the project configuration panel.
+  The notification group used is named "<i>Pyright LS notifications</i>".
+
 ### Changed
 
-* Change marketplace plugin name
+* The project is renamed from `pyright-experimental-plugin` to
+  `pyright-langserver` and the marketplace name of the plugin is now
+  "<i>Pyright Language Server</i>" instead of "<i>Pyright (Experimental)</i>".
+* UI components and messages are slightly changed.
 
 
 ## [0.1.0-poc.1] - 2024-03-17
 
 ### Added
 
-* Project initialized
+* Project initialized.
 
 
   [Unreleased]: https://github.com/InSyncWithFoo/pyright-for-pycharm/compare/v0.1.0-poc.1..HEAD